requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
COVID19Py==0.3.0
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
pycparser==2.20
python-telegram-bot==12.7
requests==2.23.0

